layout/generic/test/test_bug670058.html

<!DOCTYPE>
<html>
<head>
<title>test for bug 670058</title>
<script type="text/javascript"
  src="chrome://mochikit/content/MochiKit/packed.js"></script>
<script type="text/javascript"
  src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
<script type="text/javascript"
  src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"></script>
<link rel="stylesheet" type="text/css"
  href="chrome://mochikit/content/tests/SimpleTest/test.css" />
</head>
<body>
<textarea id="textarea" style="width: 200px; height: 100px;"></textarea>
<pre id="test">
<script class="testbody" type="text/javascript">
SimpleTest.waitForExplicitFinish();

var textarea = document.getElementById('textarea');

var x = 50;
var y = 50;

var timer = 0;

function doTest() {
  synthesizeMouse(textarea, x, y, { type: "mousedown" });

  timer = setInterval(function()
    {
      if (textarea.style.display == 'block') {
        synthesizeMouse(textarea, x++, y++, { type: "mousemove" });
        textarea.style.display = '';
      } else {
        textarea.style.display = 'block';
        synthesizeMouse(textarea, x++, y++, { type: "mousemove" });
      }
      if (x == 70) {
        ok(true, "not crashed");
        clearInterval(timer);
        SimpleTest.finish();
      }
    }, 200);
}

SimpleTest.waitForFocus(doTest, window);

</script>
</pre>
</body>
</html>
